--- src/hooks/index.js
@@ -1,3 +1,0 @@
-import useForm from "./useForm";
-
-export {useForm};
