--- src/hooks/useForm.js
@@ -1,34 +1,0 @@
-import { useState, useEffect } from 'react';
-
-const useForm = (callback, defaultValues={}) => {
-
-  const [values, setValues] = useState({});
-
-  const handleSubmit = (event) => {
-    event.preventDefault();
-    callback(values);
-  };
-
-  const handleChange = (event) => {
-    event.persist();
-
-    let { name, value } = event.target;
-    if (parseInt(value)) {
-      value = parseInt(value);
-    }
-
-    setValues(values => ({ ...values, [name]: value }));
-  };
-
-  useEffect( () => {
-    setValues( defaultValues );
-  }, [defaultValues]);
-
-  return {
-    handleChange,
-    handleSubmit,
-    values,
-  };
-};
-
-export default useForm;
